pyproject.toml
[project]
name = "pytest-firestore"
version = "0.1.0"
description = "A Pytest fixture for managing Google Cloud Firestore emulator"
authors = [
    {name = "Neale Petrillo", email = "neale.a.petrillo@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "pytest>=7.0",
    "filelock>=3.0",
]

[project.optional-dependencies]
client = ["google-cloud-firestore>=2.0"]

[project.entry-points."pytest11"]
firestore = "pytest_firestore.plugin"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = ">=0.23"
pytest-xdist = ">=3.0"
pytest-cov = ">=4.0"
google-cloud-firestore = ">=2.0"
ruff = ">=0.4"
mypy = ">=1.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
strict = true
python_version = "3.12"

[[tool.mypy.overrides]]
module = "google.cloud.*"
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
